circuit: fix extractor tests
