+++ packages/circuits/test/util.ts
@@ -116,7 +116,7 @@ export function bytesToIntChunks(
   return ints
 }
 
-export function completeArrayWithZeros(
+export function padArrayWithZeros(
   bigIntArray: bigint[],
   requiredLength: number,
 ) {
@@ -128,4 +128,46 @@ export function completeArrayWithZeros(
   }
 
   return bigIntArray
+}
+
+function bigIntsToByteArray(bigIntChunks: bigint[], bytesPerChunk = 31) {
+  const bytes: number[] = []
+
+  // Remove last chunks that are 0n
+  const cleanChunks = bigIntChunks
+    .reverse()
+    .reduce(
+      (acc: bigint[], item) =>
+        acc.length || item !== 0n ? [...acc, item] : [],
+      [],
+    )
+    .reverse()
+
+  cleanChunks.forEach((bigInt, i) => {
+    let byteCount = 0
+
+    while (bigInt > 0n) {
+      bytes.unshift(Number(bigInt & 0xffn))
+      bigInt >>= 8n
+      byteCount++
+    }
+
+    // Except for the last chunk, each chunk should be of size bytesPerChunk
+    // This will add 0s that were removed during the conversion because they are LSB
+    if (i < cleanChunks.length - 1) {
+      if (byteCount < bytesPerChunk) {
+        for (let j = 0; j < bytesPerChunk - byteCount; j++) {
+          bytes.unshift(0)
+        }
+      }
+    }
+  })
+
+  return bytes.reverse() // reverse to convert little endian to big endian
+}
+
+export function bigIntsToString(bigIntChunks: bigint[]) {
+  return bigIntsToByteArray(bigIntChunks)
+    .map(byte => String.fromCharCode(byte))
+    .join('')
 }
